# Oxidize the ConsolidateBlocks pass