… crates/accelerate/src/lib.rs
… crates/accelerate/src/two_qubit_decompose.rs
… crates/circuit/src/dag_circuit.rs
… crates/circuit/src/gate_matrix.rs
… crates/pyext/src/lib.rs
… qiskit/__init__.py
… qiskit/transpiler/passes/optimization/consolidate_blocks.py
… qiskit/transpiler/preset_passmanagers/builtin_plugins.py
… qiskit/transpiler/preset_passmanagers/common.py
… releasenotes/notes/rust-consolidation-a791a00380fc78b8.yaml
… test/python/transpiler/test_preset_passmanagers.py
… test/python/transpiler/test_solovay_kitaev.py