@mtreinish 2024-10-25
After the most recent round of changes the overall benchmarking results look like:

```
Benchmarks that have improved:                      

| Change   | Before [f2e07bc5] <main>               | After [a4229901] <consolidate-blocks>   |   Ratio | Benchmark (Parameter)                                                                           |
|----------|----------------------------------------|-----------------------------------------|---------|-------------------------------------------------------------------------------------------------|      
| -        | 1.77±0.01s                             | 1.57±0s                                 |    0.89 | utility_scale.UtilityScaleBenchmarks.time_qft('cz')                                             |      
| -        | 310±2ms                                | 272±1ms                                 |    0.88 | transpiler_levels.TranspilerLevelBenchmarks.time_quantum_volume_transpile_50_x_20(2)            |      
| -        | 49.8±0.7ms                             | 43.8±0.6ms                              |    0.88 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_qv_14_x_14(2)                        |      
| -        | 1.77±0.01s                             | 1.56±0.01s                              |    0.88 | utility_scale.UtilityScaleBenchmarks.time_qft('ecr')                                            |      
| -        | 535±1ms                                | 463±3ms                                 |    0.87 | transpiler_levels.TranspilerLevelBenchmarks.time_quantum_volume_transpile_50_x_20(3)            |      
| -        | 27.8±1ms                               | 23.4±0.4ms                              |    0.84 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm_backend_with_prop(2) |      
| -        | 76.9±2ms                               | 64.6±2ms                                |    0.84 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_qv_14_x_14(3)                        |      
| -        | 1.26±0.01s                             | 1.07±0.01s                              |    0.84 | utility_scale.UtilityScaleBenchmarks.time_qv('cx')                                              |      
| -        | 28.2±2ms                               | 23.4±0.3ms                              |    0.83 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm_backend_with_prop(3) |      
| -        | 417±3ms                                | 347±0.9ms                               |    0.83 | utility_scale.UtilityScaleBenchmarks.time_qaoa('cx')                                            |      
| -        | 222±2ms                                | 183±3ms                                 |    0.83 | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cx')                               |      
| -        | 279±4ms                                | 229±3ms                                 |    0.82 | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('ecr')                              |      
| -        | 1.72±0.01s                             | 1.38±0.02s                              |    0.81 | utility_scale.UtilityScaleBenchmarks.time_qv('cz')                                              |      
| -        | 276±2ms                                | 224±3ms                                 |    0.81 | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cz')                               |      
| -        | 1.66±0.02s                             | 1.34±0.01s                              |    0.8  | utility_scale.UtilityScaleBenchmarks.time_qv('ecr')                                             |      
| -        | 21.1±0.2ms                             | 16.8±0.1ms                              |    0.79 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm(2)                   |      
| -        | 643±4ms                                | 508±2ms                                 |    0.79 | utility_scale.UtilityScaleBenchmarks.time_qaoa('cz')                                            |      
| -        | 635±5ms                                | 500±9ms                                 |    0.79 | utility_scale.UtilityScaleBenchmarks.time_qaoa('ecr')                                           |
| -        | 24.0±0.7ms                             | 18.2±0.2ms                              |    0.76 | transpiler_levels.TranspilerLevelBenchmarks.time_transpile_from_large_qasm(3)                   |
| -        | 11.3±0.2ms                             | 3.68±0.02ms                             |    0.33 | passes.Collect2QPassBenchmarks.time_consolidate_blocks(5, 1024)                                 |
| -        | 30.5±0.2ms                             | 9.90±0.04ms                             |    0.32 | passes.Collect2QPassBenchmarks.time_consolidate_blocks(14, 1024)                                |

Benchmarks that have stayed the same:

| Change   | Before [f2e07bc5] <main>               | After [a4229901] <consolidate-blocks>   | Ratio   | Benchmark (Parameter)                                                                                  |
|----------|----------------------------------------|-----------------------------------------|---------|--------------------------------------------------------------------------------------------------------|
|          | failed                                 | failed                                  | n/a     | passes.Collect2QPassBenchmarks.time_consolidate_blocks(20, 1024)                                       |
|          | 3.71±0.02s                             | 3.79±0.1s                               | 1.02    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cx')                                                |
|          | 9.18±0.1ms                             | 9.27±0.1ms                              | 1.01    | utility_scale.UtilityScaleBenchmarks.time_bvlike('cx')                                                 |
|          | 9.21±0.05ms                            | 9.24±0.05ms                             | 1.00    | utility_scale.UtilityScaleBenchmarks.time_bvlike('cz')                                                 |
|          | 9.23±0.06ms                            | 9.25±0.07ms                             | 1.00    | utility_scale.UtilityScaleBenchmarks.time_bvlike('ecr')                                                |
|          | 4.00±0.05s                             | 3.92±0.07s                              | 0.98    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cz')                                                |
|          | 35.0±1ms                               | 34.2±0.2ms                              | 0.98    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cz')                           |
|          | 35.0±1ms                               | 34.2±0.3ms                              | 0.98    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('ecr')                          |
|          | 3.71±0.1s                              | 3.61±0.2s                               | 0.97    | utility_scale.UtilityScaleBenchmarks.time_circSU2('ecr')                                               |
|          | 166±2ms                                | 155±2ms                                 | 0.93    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cx')                                                 |
|          | 177±1ms                                | 165±0.8ms                               | 0.93    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cz')                                                 |
|          | 1.40±0.01s                             | 1.31±0s                                 | 0.93    | utility_scale.UtilityScaleBenchmarks.time_qft('cx')                                                    |
|          | 179±0.9ms                              | 165±1ms                                 | 0.92    | utility_scale.UtilityScaleBenchmarks.time_bv_100('ecr')                                                |
```
It's marginally faster than the previous run now.